updated xp to level code, recent actions menu now uses the full height of the map to display actions. Many actions still need prints to be correctly shown in the list.

+++ src/combat.rs
@@ -75,7 +75,7 @@ pub fn handle_combat(
                 // Add XP to relevant skills
                 for (skill_name, xp) in xp_gains {
                     if let Some(skill) = player.skills.get_mut(&skill_name) {
-                        skill.add_experience(xp as f32);
+                        skill.add_experience(xp as f64);
                         // Optionally, display XP gains to the player
                         // println!("{} gained {} XP.", skill_name, xp);
                     }
@@ -120,7 +120,7 @@ pub fn handle_combat(
                         let xp_gains = combat_xp_calculation(&attack_counts);
                         for (skill_name, xp) in xp_gains {
                             if let Some(skill) = player.skills.get_mut(&skill_name) {
-                                skill.add_experience(xp as f32);
+                                skill.add_experience(xp as f64);
                                 // Optionally, display XP gains to the player
                                 // println!("{} gained {} XP.", skill_name, xp);
                             }
@@ -165,7 +165,7 @@ pub fn handle_combat(
                         let xp_gains = combat_xp_calculation(&attack_counts);
                         for (skill_name, xp) in xp_gains {
                             if let Some(skill) = player.skills.get_mut(&skill_name) {
-                                skill.add_experience(xp as f32);
+                                skill.add_experience(xp as f64);
                                 // Optionally, display XP gains to the player
                                 // println!("{} gained {} XP.", skill_name, xp);
                             }
